# WebStore.TestConsole/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebStoreGB.WebAPI",
    "version": "v1"
  },
  "paths": {
...
    "/api/employees/{Id}": {
      "get": {
        "tags": [
          "EmployeesApi"
        ],
        "summary": "Получение сотрудника по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор сотрудника",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Идентификатор сотрудника",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "EmployeesApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Employee": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор сотрудника",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "description": "Имя сотрудника",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "description": "Фамилия сотрудника",
            "nullable": true
          },
          "patronymic": {
            "type": "string",
            "description": "Отчество сотрудника",
            "nullable": true
          },
          "age": {
            "type": "integer",
            "description": "Возраст сотрудника",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Информация о сотруднике"
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}